pyproject.toml
[project]
name = "rt-asset-utils"
version = "0.3.0"
description = "Asset Utils for Request Tracker"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "pandas>=1.5.0",
    "requests>=2.28.0",
    "django>=4.2,<5.0",
    "django-extensions>=3.2",
    "django-import-export>=3.3.0",
    "whitenoise>=6.5",
    "click>=8.0.0",
    "reportlab>=3.6.0",
    "qrcode>=7.3.1",
    "python-barcode>=0.13.1",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.1",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.22.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.0",
    "ldap3>=2.9.0"
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
request-tracker-utils = "request_tracker_utils:main"

[tool.setuptools.packages.find]
include = ["request_tracker_utils"]

[tool.setuptools.package-data]
"request_tracker_utils" = ["data/*.csv"]

[tool.pyright]
reportUnknownParameterType = "none"
reportUnknownArgumentType = "none"
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportMissingTypeStubs = "none"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "pyright>=1.1.0",
]
